+++ frontend/src/components/pages/variableClassificationForm.vue
@@ -65,6 +65,7 @@ export default {
 <style scoped>
     .fullWidth{
         width: 100%;
+        overflow: visible;
     }
 
 </style>