+++ frontend/src/changelog.md
@@ -3,13 +3,20 @@
 <br />
 
 ## Version 1.6.3
-October 12, 2022
+October 19, 2022
  - Fixed bug with tagged report not filtering properly since the change to multiselect.
+ - Added Dataset refresh status field to datasets, reworked several fields and possible values
+ - Fixed position of edit save floating buttons on schema page
+ - Changed alert success/error banners to be sticky to that you can see them no matter where you are on the page
+ - Fixed issue with ellipsis and visibility of description when not expanded on files and fields
+ - Files and fields field description shows full description when hovered in the collapsed view
+ - Added all fields report to dataset list
 
 
 <br />
 
 ## Table Of Contents
+- [Version 1.6.2 (October 19 2022)](#version-163)
 - [Version 1.6.2 (September 19 2022)](#version-162)
 - [Version 1.6.1 (September 16 2022)](#version-161)
 - [Version 1.6.0 (August 11 2022)](#version-160)